requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# API clients
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# YouTube API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# AI and NLP
openai==1.12.0
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
numpy==1.24.4
nltk==3.8.1
sentence-transformers==2.2.2

# Video processing and download
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
moviepy==1.0.3

# File handling
python-magic==0.4.27
filetype==1.2.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
playwright==1.40.0
psutil==5.9.6
factory-boy==3.3.0
faker==20.1.0
locust==2.17.0
gunicorn==21.2.0 